Insert SUID number to error/warning messages
- The original netCDF functionality (no MPI) provides the user with a SUID number in the warning/error message
	* This is so the user can keep track of SUID, and by extension, the spatial location, where these warnings/error occurred
- The MPI-enabled netCDF does not currently provide this information

- Now, allow MPI-enabled netCDF functionality provide the SUID number
- Create new "header" when writing warning/error messages from a compute process
- Update the datatype of variables so the datatype of "suid", when writing it out to file, is an unsigned long
	* Attribute "nSuids" within SW_MPI_DESIGNATE is now of type "size_t" (unsigned long)
	* Convert other function variables/arguments to eliminate warnings

+++ include/SW_MPI.h
@@ -162,7 +162,7 @@ void SW_MPI_report_log(
 void SW_MPI_root_find_active_sites(
     SW_DOMAIN *SW_Domain,
     unsigned long ***activeSuids,
-    int *numActiveSites,
+    size_t *numActiveSites,
     LOG_INFO *LogInfo
 );
